Enable deploying the unique contracts on mainnet, optimism, arbitrum
