+++ deploy/arbitrum/193_deploy_SimpleRewarder_FraxUSDs.ts
@@ -57,3 +57,4 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 
 export default func
 func.tags = ["SimpleRewarder"]
+func.skip = async () => true
